Modified fact return type
Modified fact return type

+++ src/Factorial.java
@@ -1,5 +1,5 @@
 public class Factorial {
-    public int fact(int num)
+    public long fact(int num)
     {
         if( num == 0 )
             return 1;
